# apps/web/app/api/ask/route.ts
import { NextResponse } from "next/server";
import { v4 as uuidv4 } from "uuid";

// Pretend-imports nga sistemet e tua ekzistuese industriale
import { eegAnalyze } from "@/lib/albi/eeg-analyze";
import { JONAEthicsGuardian } from "@/lib/jona/guardian";
import { AlbaCore } from "@/lib/alba/core";

const alba = new AlbaCore(32);
const jona = new JONAEthicsGuardian({ JONA_MAX_CONCURRENCY: 4 });
await jona.initialize();

// Simulon një klasë runtime për event-trace (industrial observability)
class Telemetry {
  id: string;
  logs: string[] = [];
  start = Date.now();
  constructor() {
    this.id = `REQ-${uuidv4().split("-")[0]}`;
    this.log("initialized");
  }
  log(msg: string) {
    const ts = new Date().toISOString();
    this.logs.push(`[${ts}] ${msg}`);
  }
  finish() {
    const elapsed = Date.now() - this.start;
    this.log(`completed in ${elapsed}ms`);
    return elapsed;
  }
}

export async function POST(request: Request) {
  const telemetry = new Telemetry();

  try {
    const body = await request.json();
    const message: string = body?.message ?? "";
    const userContext = body?.context ?? { user: "anonymous" };

    telemetry.log(`Received message: "${message}" from ${userContext.user}`);

    // Step 1: EEG / ALBI Neural Analysis
    telemetry.log("Analyzing EEG / neural pattern input...");
    if (!body.filePath) {
      return NextResponse.json({ error: "filePath is required for EEG analysis" }, { status: 400 });
    }
    const eeg = await eegAnalyze({ PYTHON: "python3" }, body.filePath);

    // Step 2: ALBA Stream Integration
    telemetry.log("Querying ALBA Industrial Streams...");
    const albaStatus = alba.getStatus();
    const albaPattern = alba.analyzePatterns();
    const albaAlerts = alba.getAlerts();

    // Step 3: JONA Ethical Review
    telemetry.log("Running ethical evaluation via JONA...");
    const actionProposal = {
      id: uuidv4(),
      description: "Neural command processing from mobile interface",
      origin: "Clisonix Cloud",
    };
    const context = { user: userContext.user, environment: "industrial", signal: eeg?.bands };
    const ethics = await jona.evaluateEthicalAction(actionProposal, context);

    // Step 4: Result Synthesis
    telemetry.log("Synthesizing final report and confidence metrics...");
    const time = telemetry.finish();

    const response = {
      request_id: telemetry.id,
      modules_used: ["ALBI", "ALBA", "JONA"],
      eeg: {
        ok: eeg.ok,
        dominant_hz: eeg.dominant_hz,
        bands: eeg.bands,
      },
      alba: {
        status: albaStatus.status,
        security_level: albaStatus.securityLevel,
        trend: albaPattern.trend,
        alerts: albaAlerts.length,
      },
      jona: {
        approval: ethics.overallApproval,
        confidence: ethics.confidence,
        warnings: ethics.warnings,
        recommendations: ethics.recommendations,
      },
      runtime: {
        processing_ms: time,
        system_load: albaStatus.cpuLoad,
        memory_usage: albaStatus.memoryUsage,
      },
      logs: telemetry.logs,
      timestamp: new Date().toISOString(),
    };

    return NextResponse.json(response, { status: 200 });
  } catch (err: unknown) {
    telemetry.log(`❌ Error: ${err instanceof Error ? err.message : 'Unknown error'}`);
    const errorResp = {
      error: "Processing failure in ALBA/JONA pipeline",
      details: err instanceof Error ? err.message : 'Unknown error',
      logs: telemetry.logs,
      timestamp: new Date().toISOString(),
    };
    return NextResponse.json(errorResp, { status: 500 });
  }
}