# openapi.yaml
openapi: 3.1.0
info:
  title: Clisonix Cloud API - Part of Webultrathinking Euroweb
  version: '1.0.0'
  description: |
    Clisonix Cloud API – Enterprise-grade neuro-audio, EEG processing, brain-sync engine,
    billing integration, and real-time system observability.
    
    **Key Features:**
    - Neural harmonic synchronization engine
    - EEG signal processing and analysis
    - Brain-adaptive music generation
    - Real-time data streaming (ALBA)
    - Multi-channel payment integration
    - Industrial-grade system monitoring
    
    **Security:** All endpoints require Bearer JWT authentication.
  contact:
    name: Clisonix Support
    url: https://clisonix.cloud/support
    email: clisonix@pm.me
  license:
    name: Proprietary
    url: https://clisonix.cloud/license
  x-logo:
    url: https://clisonix.cloud/logo.png
servers:
  - url: http://localhost:8000
    description: Local Development
    variables: {}
  - url: https://api.clisonix.com
    description: Production
    variables: {}
  - url: https://sandbox.clisonix.cloud
    description: Sandbox Testing
    variables: {}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Bearer token. Format: Authorization: Bearer <token>'
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Alternative API key authentication
    oauth2:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /auth/token
          scopes:
            read: Read access to resources
            write: Write access to resources
            admin: Administrative access
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - AUTHENTICATION_FAILED
                - UNAUTHORIZED
                - NOT_FOUND
                - VALIDATION_ERROR
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
            message:
              type: string
            details:
              type: string
              nullable: true
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
...
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          format: password
          description: User password
    AuthLoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
        token_type:
          type: string
          default: bearer
          description: Type of token
        expires_in:
          type: integer
          description: Token expiration time in seconds
          example: 3600
        refresh_token:
          type: string
          description: Token for refreshing access
        api_key:
          type: string
          nullable: true
          description: Optional API key for this user
...
    AskRequest:
      type: object
      required:
        - question
      properties:
        question:
          type: string
          minLength: 1
          maxLength: 2000
          description: Question or query for the AI assistant
        context:
          type: string
          nullable: true
          maxLength: 5000
          description: Optional context to help answer the question
        include_details:
          type: boolean
          default: true
          description: Include detailed analysis in response
    AskResponse:
      type: object
      properties:
        answer:
          type: string
          description: AI-generated answer
        timestamp:
          type: string
          format: date-time
        modules_used:
          type: array
          items:
            type: string
          example: ['brain', 'cortex', 'harmony']
        processing_time_ms:
          type: integer
          minimum: 0
        details:
          type: object
          additionalProperties: true
          nullable: true
...
  responses:
    UnauthorizedError:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: string
            format: date-time
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
  - bearerAuth: []
  - apiKey: []
paths:
  /auth/login:
    post:
      summary: User Login
      tags:
        - Core
      description: Authenticate user with email & password and return JWT + refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/ask:
    post:
      summary: Ask AI assistant
      tags:
        - Brain
      description: Query the AI assistant with optional context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskRequest'
      responses:
        '200':
          description: Question answered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...